Reworked server structure
