+++ CMakeLists.txt
@@ -105,7 +105,7 @@ target_include_directories(mvp PRIVATE
 )
 
 # Create mvp server
-add_executable(mvp_server src/mvp/mvpserver.cpp)
+add_executable(mvp_server src/mvp/MvpServer.cpp)
 target_link_libraries(mvp_server engine)
 target_include_directories(mvp_server PRIVATE
     "src/mvp/"
